requirements.txt
django>=3.2
djangorestframework
djangorestframework-csv
drf-excel
